pyproject.toml
[project]
name = "agent-see"
version = "0.1.0"
description = "Convert any SaaS application into an agent-optimized interface"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "playwright>=1.40.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "openapi-spec-validator>=0.7.0",
    "rich>=13.0.0",
    "anthropic>=0.40.0",
    "tiktoken>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
agent-see = "agent_see.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_see"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
